requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
email-validator
bcrypt
Pillow